Refactor bucket and point related tests outside CStatisticsTest
