+++ server/include/unittest/dxl/statistics/CStatisticsTest.h
@@ -63,50 +63,6 @@ namespace gpnaucrates
 				const CHAR *szMergedFile;
 			};
 
-			// test case for intersection of buckets
-			struct SBucketsIntersectTestElem
-			{
-				 // lower bound of bucket 1
-				INT m_iLb1;
-
-				 // upper bound of bucket 1
-				INT m_iUb1;
-
-				// is lower bound of bucket 1 closed
-				BOOL m_fLb1Closed;
-
-				// is upper bound of bucket 1 closed
-				BOOL m_fUb1Closed;
-
-				// lower bound of bucket 2
-				INT m_iLb2;
-
-				// upper bound of bucket 2
-				INT m_iUb2;
-
-				// is lower bound of bucket 2 closed
-				BOOL m_fLb2Closed;
-
-				// is upper bound of bucket 2 closed
-				BOOL m_fUb2Closed;
-
-				// result of the bucket intersect test
-				BOOL fIntersect;
-
-				// lower bound of output bucket
-				INT m_iLbOutput;
-
-				// upper bound of output bucket
-				INT m_iUbOutput;
-
-				// is lower bound of output bucket closed
-				BOOL m_fLbOutputClosed;
-
-				// is upper bound of output bucket closed
-				BOOL m_fUbOutputClosed;
-
-			}; // SBucketsIntersectTestElem
-
 			// test case for disjunctive filter evaluation
 			struct SStatsFilterSTestCase
 			{
@@ -212,10 +168,6 @@ namespace gpnaucrates
 				const CWStringConst &strColA,
 				const CWStringConst &strColB
 				);
-
-			// do the bucket boundaries match
-			static
-			BOOL FMatchBucketBoundary(CBucket *pbucket1, CBucket *pbucket2);
 
 		public:
 
@@ -332,23 +284,9 @@ namespace gpnaucrates
 			static
 			GPOS_RESULT EresUnittest();
 
-			// point related tests
-			static
-			GPOS_RESULT EresUnittest_CPointInt4();
-
-			static
-			GPOS_RESULT EresUnittest_CPointBool();
-
 			// join of histograms with NDVRemain information
 			static
 			GPOS_RESULT EresUnittest_JoinNDVRemain();
-
-			// bucket basic tests
-			static
-			GPOS_RESULT EresUnittest_CBucketInt4();
-
-			static
-			GPOS_RESULT EresUnittest_CBucketBool();
 
 			// join buckets tests
 			static
@@ -357,18 +295,6 @@ namespace gpnaucrates
 			// union all tests
 			static
 			GPOS_RESULT EresUnittest_UnionAll();
-
-			// bucket intersect
-			static
-			GPOS_RESULT EresUnittest_CBucketIntersect();
-
-			// bucket scaling tests
-			static
-			GPOS_RESULT EresUnittest_CBucketScale();
-
-			// bucket difference tests
-			static
-			GPOS_RESULT EresUnittest_CBucketDifference();
 
 			// histogram basic tests
 			static
